requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
asgiref==3.8.1
asttokens
async-timeout==4.0.3
attrs==25.3.0
auth0-python==4.9.0
Authlib==1.3.1
backcall==0.2.0
backoff==2.2.1
backports.tarfile==1.2.0
banks==2.1.2
bcrypt==4.3.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
build==1.2.2.post1
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
chroma-hnswlib==0.7.6
chromadb==0.5.23
cleo==2.1.0
click==8.1.8
cohere==5.15.0
colorama
coloredlogs==15.0.1
comm
crashtest==0.4.1
crewai==0.119.0
crewai-tools==0.44.0
cryptography==44.0.3
dataclasses-json==0.6.7
debugpy
decorator
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
dirtyjson==1.0.8
distlib==0.3.9
distro==1.9.0
docker==7.1.0
docopt==0.6.2
docstring_parser==0.16
dulwich==0.22.8
durationpy==0.9
embedchain==0.1.128
et_xmlfile==2.0.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
executing
faiss-cpu==1.11.0
fastapi==0.115.9
fastavro==1.10.0
fastjsonschema==2.21.1
filelock==3.18.0
filetype==1.2.0
findpython==0.6.3
Flask==3.1.0
flask-cors==5.0.1
flatbuffers==25.2.10
frozenlist==1.6.0
fsspec==2025.3.2
gensim==4.3.3
google-auth==2.40.1
googleapis-common-protos==1.70.0
gptcache==0.1.44
greenlet==3.2.2
griffe==1.7.3
grpcio==1.71.0
grpcio-health-checking==1.71.0
grpcio-tools==1.71.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.31.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib_metadata
importlib_resources==6.5.2
installer==0.7.0
instructor==1.8.1
ipython==8.12.3
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi
Jinja2==3.1.6
jiter==0.8.2
joblib==1.5.0
json5==0.12.0
json_repair==0.44.1
jsonpatch==1.33
jsonpickle==4.0.5
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
keyring==25.6.0
kubernetes==32.0.1
lancedb==0.22.0
langsmith==0.3.42
litellm==1.68.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline
mdurl==0.1.2
mem0ai==0.1.98
mistralai==1.7.0
mistune==3.1.3
mmh3==5.1.0
monotonic==1.6
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest_asyncio
networkx==3.4.2
nltk==3.9.1
nodeenv==1.9.1
notion-client==2.3.0
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.21.1
openai==1.75.0
openpyxl==3.1.5
opentelemetry-api==1.33.0
opentelemetry-exporter-otlp-proto-common==1.33.0
opentelemetry-exporter-otlp-proto-grpc==1.33.0
opentelemetry-exporter-otlp-proto-http==1.33.0
opentelemetry-instrumentation==0.54b0
opentelemetry-instrumentation-asgi==0.54b0
opentelemetry-instrumentation-fastapi==0.54b0
opentelemetry-proto==1.33.0
opentelemetry-sdk==1.33.0
opentelemetry-semantic-conventions==0.54b0
opentelemetry-util-http==0.54b0
orjson==3.10.18
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso
pathtools==0.1.2
pbs-installer==2025.4.9
pdfminer.six==20250327
pdfplumber==0.11.6
pickleshare
pillow==11.2.1
pkginfo==1.12.1.2
platformdirs
poetry==2.1.3
poetry-core==2.1.3
portalocker==2.10.1
posthog==3.25.0
prompt_toolkit
propcache==0.3.1
protobuf==5.29.4
psutil
ptyprocess==0.7.0
pure_eval
py==1.11.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments
PyJWT==2.10.1
pypdf==5.4.0
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pyright==1.1.400
pysbd==0.3.4
python-dateutil
python-dotenv==1.1.0
pytube==15.0.0
pytz==2024.2
pyvis==0.3.2
PyYAML==6.0.2
pyzmq
qdrant-client==1.14.2
RapidFuzz==3.13.0
redis==3.5.3
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.24.0
rsa==4.9.1
safetensors==0.5.3
schema==0.7.7
scikit-learn==1.6.1
scipy==1.13.1
shellingham==1.5.4
six
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
sqlparse==0.5.3
stack_data
starlette==0.45.3
striprtf==0.0.26
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
textblob==0.19.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.20.3
toml==0.10.2
tomli==2.2.1
tomli_w==1.2.0
tomlkit==0.13.2
tornado
tqdm==4.67.1
traitlets
trove-classifiers==2025.5.9.12
txaio==23.1.1
txtai==8.5.0
typer==0.15.3
types-requests==2.32.0.20250328
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions
tzdata==2025.2
urllib3==2.4.0
uv==0.7.3
uvicorn==0.34.2
validators==0.34.0
vine==5.1.0
virtualenv==20.31.2
watchdog==0.10.4
watchfiles==1.0.5
wcwidth
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
yarg==0.1.9
yarl==1.20.0
zipp
zope.interface==7.2
zstandard==0.23.0
